# gateway/api-gateway/openapi.yaml
...
openapi: 3.0.3
info:
  title: Food Delivery API
  version: 1.0.0
  description: |
    Every request goes through the **api-gateway**, which routes it to the owning microservice.

    **To call protected endpoints:** run `POST /login` (e.g. `jamie@example.com` / `password123`),
    copy the `token` from the response, click **Authorize** and paste it.
    For the **Owner** endpoints, log in as a restaurant owner instead
    (e.g. `sam@mariospizzeria.com` / `password123`).

    **Order flow (event-driven over RabbitMQ):** `POST /orders` publishes `order.created` →
    payment-service charges it and publishes `payment.succeeded` or `payment.failed` →
    order-service moves the order to `CONFIRMED` or `CANCELLED` → restaurant-service publishes
    `restaurant.order_received`. The flow is asynchronous, so poll `GET /orders/{id}` for the
    outcome. Orders over 100.00 are declined by the simulated payment provider, so the failure
    path can be tested too.

    **Restaurant side:** a paid (`CONFIRMED`) order shows up in `GET /owner/orders`. The owner
    accepts it (`PREPARING`, publishes `order.preparing`), marks it ready (`READY`, publishes
    `order.ready`) or rejects it (`CANCELLED` and refunded). Customer and owner both get every
    status change pushed over the notification websocket.
servers:
  - url: /
    description: This api-gateway
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List the logged-in customer's orders
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Orders, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Orders]
      summary: Place an order
      description: >
        Prices are looked up from restaurant-service. The order starts as
        `PLACED`; payment happens asynchronously over the event bus.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrder' }
      responses:
        '201':
          description: Order placed and `order.created` published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: A menu item is unknown or unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502': { $ref: '#/components/responses/Upstream' }
        '503':
          description: Event bus unreachable; the order was not placed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get one of your orders
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order
      description: >
        Allowed while `PLACED` or `CONFIRMED`. Publishes `order.cancelled`;
        payment-service then refunds a payment that already succeeded.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /owner/orders:
    get:
      tags: [Owner orders]
      summary: List paid orders for your restaurants
      description: Oldest first. Orders still awaiting payment, or whose payment failed, are never shown.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          description: Comma-separated statuses, e.g. `CONFIRMED,PREPARING,READY`
          schema: { type: string, example: 'CONFIRMED,PREPARING,READY' }
        - name: restaurantId
          in: query
          description: Only this restaurant's orders
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /owner/orders/{id}:
    get:
      tags: [Owner orders]
      summary: Get one of your restaurants' orders
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /owner/orders/{id}/accept:
    post:
      tags: [Owner orders]
      summary: Accept an order and start preparing it
      description: '`CONFIRMED` → `PREPARING`. Publishes `order.preparing`.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200': { $ref: '#/components/responses/OwnerOrder' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /owner/orders/{id}/reject:
    post:
      tags: [Owner orders]
      summary: Reject an order
      description: >
        Allowed while `CONFIRMED` or `PREPARING`. The order is `CANCELLED`,
        `order.cancelled` is published and payment-service refunds it.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 200 }
            example: { reason: 'Out of mozzarella' }
      responses:
        '200': { $ref: '#/components/responses/OwnerOrder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflicts with the current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Logged in, but not as a restaurant owner
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    OwnerOrder:
      description: The order in its new status
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Order' }
    Upstream:
      description: A downstream service or the event bus is unreachable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    CreateOrder:
      type: object
      required: [restaurantId, items]
      properties:
        restaurantId: { type: string, format: uuid }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [menuItemId, quantity]
            properties:
              menuItemId: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
      example:
        restaurantId: 33333333-3333-3333-3333-333333333331
        items:
          - { menuItemId: 44444444-4444-4444-4444-444444444441, quantity: 2 }
          - { menuItemId: 44444444-4444-4444-4444-444444444442, quantity: 1 }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        status:
          type: string
          enum: [PLACED, CONFIRMED, PREPARING, READY, PICKED_UP, DELIVERED, COMPLETED, CANCELLED]
        totalAmount: { type: string, example: '29.5' }
        confirmedAt: { type: string, format: date-time, nullable: true, description: 'When payment succeeded' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lines:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              menuItemId: { type: string, format: uuid }
              name: { type: string }
              unitPrice: { type: string, example: '12.5' }
              quantity: { type: integer }
...